trajectory initializer class
